+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
+### Changed
+- Exposed interface names are now correctly prefixed. ([craftcms/gatsby-source-craft#58](https://github.com/craftcms/gatsby-source-craft/issues/58))
+
 ### Fixed
-- Fixed an error in the SQL query when fetching updated elements. ([#57](https://github.com/craftcms/gatsby-source-craft/issues/57))
+- Fixed an error in the SQL query when fetching updated elements. ([craftcms/gatsby-source-craft#57](https://github.com/craftcms/gatsby-source-craft/issues/57))
 
 ## 1.0.8 - 2021-11-25